refactor: add missing namespaces and rename mispelling directories
