+++ LMS.DataAccess/src/Console/UserMenu/PatronMenu/PatronMenu.cs
@@ -1,5 +1,9 @@
 using LMS.DataAccess.BookSystem.Concretes;
 using LMS.DataAccess.BorrowSystem;
+using LMS.DataAccess.Console.UserMenu.PatronMenu.Borrow;
+using LMS.DataAccess.Console.UserMenu.PatronMenu.Search;
+using LMS.DataAccess.Console.Utils.ItemsList;
+using LMS.DataAccess.Console.Utils.ItemsList.Concretes;
 using LMS.DataAccess.FineSystem;
 using LMS.DataAccess.UserSystem.Concretes;
 using Spectre.Console;
