+++ LMS.DataAccess/src/Console/UserMenu/PatronMenu/Search/OptionSearch.cs
@@ -2,7 +2,7 @@ using LMS.DataAccess.Console.UserMenu.PatronMenu.Search.Interfaces;
 
 namespace LMS.DataAccess.Console.UserMenu.PatronMenu.Search;
 
-public class SearchOption {
+public class OptionSearch {
     SearchInput? _search;
     public void SetSearchOption(SearchInput _search){
         this._search = _search;
