--- LMS.DataAccess/src/Console/UserMenu/StaffMenu/Book/Interfaces/BookInput.cs
@@ -1,6 +1,0 @@
-namespace LMS.DataAccess.Console.UserMenu.StaffMenu.Book.Interfaces;
-
-public interface BookInput
-{
-    public void BookOption();
-}
