+++ LMS.DataAccess/src/Console/UserMenu/StaffMenu/Books/Interfaces/BookInput.cs
@@ -1,0 +1,6 @@
+namespace LMS.DataAccess.Console.UserMenu.StaffMenu.Books.Interfaces;
+
+public interface BookInput
+{
+    public void BookOption();
+}
