+++ LMS.DataAccess/src/Console/UserMenu/StaffMenu/StaffMenu.cs
@@ -1,5 +1,9 @@
 using LMS.DataAccess.BookSystem.Concretes;
 using LMS.DataAccess.BorrowSystem;
+using LMS.DataAccess.Console.UserMenu.StaffMenu.Books;
+using LMS.DataAccess.Console.UserMenu.StaffMenu.Report;
+using LMS.DataAccess.Console.Utils.ItemsList;
+using LMS.DataAccess.Console.Utils.ItemsList.Concretes;
 using LMS.DataAccess.FineSystem;
 using LMS.DataAccess.UserSystem;
 using LMS.DataAccess.UserSystem.Concretes;
